Creating plugboard for Enigma Machine.
The plugboard is the first step in the Enigma machine, where the
operator may connect two letters and swap their values when typing on
the keyboard. For instance, if letters `A` and `B` were plugged, whenever the operator types `A`, the machine will output `B` and vice-versa.

Letters that haven't been plugged to another one will output their usual
value.