First simplified but functional Enigma class.
The current Enigma has only three components:
- Plugboard
- One rotor
- Reflector

An encoded letter first goes through the plugboard, then through the
rotor and reaches the reflector. At this point, it returns through the
rotor and plugboard to become the output encoded letter.

So far, the components substitution tables are hardcoded, which should
be evolver in the future. The rotor doesn't step yet and two more rotors
must be added, with the functionality to define their turnover points
and initial positions.

Nice, we're getting there.